+++ src/states/game/StartMenu.lua
@@ -7,6 +7,7 @@ function StartMenu:init()
     self.newButton = Button(BUTTON_PARAMS['New'])
     self.settingsButton = Button(BUTTON_PARAMS['StartSettings'])
     self.background = StartMenuBackground()
+    self.popup = PopupWindow(POPUP_WINDOW_CONFIG)
 
     if love.filesystem.getInfo(SAVE_FILE) then
         self.playButton:enable()
@@ -21,6 +22,7 @@ function StartMenu:init()
     gStateStack:push(self.background)
     gStateStack:push(self.playButton)
     gStateStack:push(self.newButton)
+    gStateStack:push(self.popup)
     gStateStack:push(self.settingsButton)
 
     -- Start music when the window opens